src/requirements.txt
asgiref==3.2.9
Django==3.0.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
Pillow==7.1.2
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
